reorganize tools and metadata. Update file io to maintain metadata format.
