+++ holopy/core/metadata.py
@@ -22,27 +22,95 @@ Classes for defining metadata about experimental or calculated results.
 
 """
 
-
 import numpy as np
 import xarray as xr
-from xarray.ufuncs import sqrt, arctan2
+from xarray.ufuncs import sqrt,arctan2
 from warnings import warn
-from copy import copy
-from .tools import _ensure_pair, _ensure_array, is_none, flat, from_flat, updated
-
+from .utils import _ensure_pair, _ensure_array, is_none, updated
 
 vector = 'vector'
+
+def Image(arr, spacing=None, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=(0, 0, 1)):
+    if np.isscalar(spacing):
+        spacing = np.repeat(spacing, 2)
+
+    out = xr.DataArray(arr, dims=['x', 'y'], coords=make_coords(arr.shape, spacing))
+    return update_metadata(out, medium_index, illum_wavelen, illum_polarization,normals)    
+
+def ImageSchema(shape, spacing, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=(0, 0, 1)):
+    if np.isscalar(shape):
+        shape = np.repeat(shape, 2)
+
+    d = np.zeros(shape)
+    return Image(d, spacing, medium_index, illum_wavelen, illum_polarization, normals)
+
+def update_metadata(a, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=None):
+    attrlist = {'medium_index': medium_index, 'illum_wavelen': illum_wavelen, 'illum_polarization': to_vector(illum_polarization), 'normals': to_vector(normals)}
+    b = a.copy()
+    b.attrs = updated(b.attrs, attrlist)
+
+    for attr in attrlist:
+        if not hasattr(b, attr):
+            b.attrs[attr] = None
+
+    if is_none(b.normals):
+        b.attrs['normals'] = to_vector((0,0,1))
+
+    return b
+
+def copy_metadata(old, new, do_coords=True):
+    def find_and_rename(oldkey, oldval):
+        for newkey, newval in new.coords.items():
+            if np.array_equal(oldval.values, newval.values):
+                return new.rename({newkey: oldkey})
+            raise ValueError("Coordinate {} does not appear to have a coresponding coordinate in {}".format(oldkey, new))
+
+    if hasattr(old, 'attrs') and hasattr(old, 'name') and hasattr(old, 'coords'):
+        if not hasattr(new,'coords'):
+            #new is a numpy array, not xarray
+            new=xr.DataArray(new, dims=['x', 'y'])
+        new.attrs = old.attrs
+        new.name = old.name
+        if hasattr(old, 'z') and not hasattr(new, 'z'):
+            new.coords['z'] = old.coords['z']
+        if hasattr(old, 'flat') and hasattr(new, 'flat'):
+            new['flat'] = old['flat']
+        if do_coords:
+            for key, val in old.coords.items():
+                if key not in new.coords:
+                    new = find_and_rename(key, val)
+    return new
 
 def to_vector(c):
     if c is None:
         return c
-    if hasattr(c, 'vector'):
+    if hasattr(c, vector):
         return c
     c = np.array(c)
     if c.shape == (2,):
         c = np.append(c, 0)
 
     return xr.DataArray(c, coords={vector: ['x', 'y', 'z']})
+
+def flat(a, keep_xy=True):
+    if hasattr(a, 'flat'):
+        # TODO handle case where we have flat but not xyz
+        return a
+    if hasattr(a, 'x') and hasattr(a, 'y') and keep_xy:
+        a['x_orig'] = a.x
+        a['y_orig'] = a.y
+        # TODO: remove *_orig coords from a or avoid adding them
+        f = a.stack(flat=a.dims)
+        del a['x_orig']
+        del a['y_orig']
+        return f.rename({'x_orig': 'x', 'y_orig': 'y'})
+    else:
+        return a.stack(flat=a.dims)
+
+def from_flat(a):
+    if hasattr(a, 'flat'):
+        return a.unstack('flat')
+    return a
 
 def to_spherical(a, origin, wavevec=None, include_r=True):
     xo, yo, zo = origin
@@ -65,92 +133,19 @@ def to_spherical(a, origin, wavevec=None, include_r=True):
 def r_theta_phi_flat(a, origin):
     f = flat(to_spherical(a, origin))
     return f
-    return np.vstack((f.r, f.theta, f.phi)).T
 
 def kr_theta_phi_flat(a, origin, wavevec=None):
     if wavevec is None:
-        wavevec = get_wavevec(a)
+        wavevec = 2*np.pi/(a.illum_wavelen/a.medium_index)
     return flat(to_spherical(a, origin, wavevec))
-    pos = r_theta_phi_flat(a, origin)
-    pos['r'] *= wavevec
-    return pos
 
 def theta_phi_flat(a, origin=None):
     if hasattr(a, 'theta') and hasattr(a, 'phi'):
         return a
     return flat(to_spherical(a, origin))
 
-def make_coords(shape, spacing, z=0):
-    if np.isscalar(shape):
-        shape = np.repeat(shape, 2)
-    if np.isscalar(spacing):
-        spacing = np.repeat(spacing, 2)
-    return {'x': np.arange(shape[0])*spacing[0], 'y': np.arange(shape[1])*spacing[1], 'z': 0}
-
-def update_metadata(a, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=None):
-    attrlist = {'medium_index': medium_index, 'illum_wavelen': illum_wavelen, 'illum_polarization': to_vector(illum_polarization), 'normals': to_vector(normals)}
-    b = copy(a)    
-    b.attrs = updated(b.attrs, attrlist)
-
-    for attr in attrlist:
-        if not hasattr(b, attr):
-            b.attrs[attr]=None
-
-    if is_none(b.normals):
-        b.attrs['normals']=to_vector((0,0,1))
-
-    return b
-
-def angles_list(theta, phi, medium_index, illum_wavelen, illum_polarization, normals=(0, 0, 1)):
-    # This is a hack that gets the data into a format that we can use
-    # elsewhere, but feels like an abuse of xarray, it would be nice to replace this with something more ideomatic
-
-    theta = _ensure_array(theta)
-    phi = _ensure_array(phi)
-    if len(theta) == 1:
-        theta = np.repeat(theta,len(phi))
-    elif len(phi) == 1:
-        phi = np.repeat(phi,len(theta))
-
-    out = xr.DataArray(np.zeros(len(theta)), dims=['point'], attrs={'theta':theta, 'phi':phi})
-    return update_metadata(out, medium_index, illum_wavelen, illum_polarization, normals)
-
-def ImageSchema(shape, spacing, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=(0, 0, 1)):
-    if np.isscalar(shape):
-        shape = np.repeat(shape, 2)
-
-    d = np.zeros(shape)
-    return Image(d, spacing, medium_index, illum_wavelen, illum_polarization, normals)
-
-def Image(arr, spacing=None, medium_index=None, illum_wavelen=None, illum_polarization=None, normals=(0, 0, 1)):
-    if np.isscalar(spacing):
-        spacing = np.repeat(spacing, 2)
-
-    out = xr.DataArray(arr, dims=['x', 'y'], coords=make_coords(arr.shape, spacing))
-    return update_metadata(out, medium_index, illum_wavelen, illum_polarization,normals)    
-
-def get_med_wavelen(a):
-    return a.illum_wavelen/a.medium_index
-
-def get_wavevec(a):
-    return 2*np.pi/get_med_wavelen(a)
-
-def optical_parameters(schema=None, **kwargs):
-    d = {}
-    for par, val in kwargs.items():
-        if val is None:
-            d[par] = getattr(schema, par)
-        else:
-            d[par] = val
-    r = {}
-    if 'illum_wavelen' in d and 'medium_index' in d:
-        r['medium_wavevec'] = 2*np.pi/(d['illum_wavelen']/d['medium_index'])
-    if 'medium_index' in d:
-        r['medium_index'] = d['medium_index']
-    if 'illum_polarization' in d:
-        r['illum_polarization'] = to_vector(d['illum_polarization'])
-
-    return r
+def get_values(a):
+    return getattr(a, 'values', a)
 
 def get_spacing(im):
     xspacing = np.diff(im.x)
@@ -158,3 +153,10 @@ def get_spacing(im):
     if not np.allclose(xspacing[0], xspacing) and np.allclose(yspacing[0], yspacing):
         raise ValueError("array has nonuniform spacing, can't determine a single spacing")
     return np.array((xspacing[0], yspacing[0]))
+
+def make_coords(shape, spacing, z=0):
+    if np.isscalar(shape):
+        shape = np.repeat(shape, 2)
+    if np.isscalar(spacing):
+        spacing = np.repeat(spacing, 2)
+    return {'x': np.arange(shape[0])*spacing[0], 'y': np.arange(shape[1])*spacing[1], 'z': 0}
