+++ holopy/core/process/__init__.py
@@ -27,10 +27,6 @@ reconstructions.
 '''
 
 
-from .img_proc import normalize, detrend, zero_filter, subimage, add_noise, fft, ifft
+from .img_proc import normalize, detrend, zero_filter, subimage, add_noise, simulate_noise
+from .fourier import fft, ifft
 from .centerfinder import center_find, centered_subimage, hough, image_gradient
-from .math import cartesian_distance, rotate_points, rotation_matrix,simulate_noise
-from .utilities import _ensure_array, ensure_listlike, _ensure_pair, mkdir_p
-from .utilities import dict_without, is_none, updated,  get_values
-from .utilities import copy_metadata, flat, from_flat, make_subset_data
-from .math import chisq, rsq
