+++ holopy/core/utils.py
@@ -26,7 +26,6 @@ import os
 import shutil
 import errno
 import numpy as np
-import xarray as xr
 from copy import copy
 import itertools
 
@@ -125,61 +124,3 @@ def updated(d, update={}, filter_none=True, **kwargs):
             d[key] = val
 
     return d
-
-def copy_metadata(old, new, do_coords=True):
-    def find_and_rename(oldkey, oldval):
-        for newkey, newval in new.coords.items():
-            if np.array_equal(oldval.values, newval.values):
-                return new.rename({newkey: oldkey})
-            raise ValueError("Coordinate {} does not appear to have a coresponding coordinate in {}".format(oldkey, new))
-
-    if hasattr(old, 'attrs') and hasattr(old, 'name') and hasattr(old, 'coords'):
-        if not hasattr(new,'coords'):
-            #new is a numpy array, not xarray
-            new=xr.DataArray(new, dims=['x', 'y'])
-        new.attrs = old.attrs
-        new.name = old.name
-        if hasattr(old, 'z') and not hasattr(new, 'z'):
-            new.coords['z'] = old.coords['z']
-        if hasattr(old, 'flat') and hasattr(new, 'flat'):
-            new['flat'] = old['flat']
-        if do_coords:
-            for key, val in old.coords.items():
-                if key not in new.coords:
-                    new = find_and_rename(key, val)
-    return new
-
-def get_values(a):
-    return getattr(a, 'values', a)
-
-def flat(a, keep_xy=True):
-    if hasattr(a, 'flat'):
-        # TODO handle case where we have flat but not xyz
-        return a
-    if hasattr(a, 'x') and hasattr(a, 'y') and keep_xy:
-        a['x_orig'] = a.x
-        a['y_orig'] = a.y
-        # TODO: remove *_orig coords from a or avoid adding them
-        f = a.stack(flat=a.dims)
-        del a['x_orig']
-        del a['y_orig']
-        return f.rename({'x_orig': 'x', 'y_orig': 'y'})
-    else:
-        return a.stack(flat=a.dims)
-
-def from_flat(a):
-    if hasattr(a, 'flat'):
-        return a.unstack('flat')
-    return a
-
-def make_subset_data(data, random_subset, return_selection=False):
-    if random_subset is None:
-        return data
-    n_sel = int(np.ceil(data.size*random_subset))
-    selection = np.random.choice(data.size, n_sel, replace=False)
-    subset = flat(data)[selection]
-    subset = copy_metadata(data, subset, do_coords=False)
-    if return_selection:
-        return subset, selection
-    else:
-        return subset
