+++ holopy/fitting/fit_series.py
@@ -29,7 +29,8 @@ import warnings
 import os
 import types
 import numpy as np
-from holopy.core.tools import normalize, subimage, mkdir_p
+from holopy.core.utils import mkdir_p
+from holopy.core.process import normalize, subimage
 from holopy.core import Image
 from holopy.core.io import load, save
 from holopy.fitting import fit
