+++ holopy/scattering/scatterer/csg.py
@@ -24,7 +24,7 @@ the DDA th
 '''
 from holopy.scattering.errors import InvalidScatterer
 from holopy.scattering.scatterer import Scatterer
-from holopy.core.tools.math import rotate_points
+from holopy.core.math import rotate_points
 
 import numpy as np
 from numpy import logical_and, logical_not, logical_or
